+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 main
 ----
+* Update GainTable to read different receptor frames as inputs ([MR22](https://gitlab.com/ska-telescope/sdp/ska-sdp-datamodels/-/merge_requests/22))
 * get_direction_time_location and calculate_visibility_hourangles allow for user-defined time inputs ([MR20](https://gitlab.com/ska-telescope/sdp/ska-sdp-datamodels/-/merge_requests/20))
 * Bug fix in PointingTable.pointingtable_acc.nrec ([MR20](https://gitlab.com/ska-telescope/sdp/ska-sdp-datamodels/-/merge_requests/20))
 * Bug fix when simulating with more than 2045 frequency channels (([MR12](https://gitlab.com/ska-telescope/sdp/ska-sdp-datamodels/-/merge_requests/12)))